+++ packages/core/icon/src/components/__tests__/unit/indexSpec.js
@@ -1,15 +1,13 @@
 // @flow
 import React, { Component } from 'react';
-import { mount, shallow, render } from 'enzyme';
+import { mount, shallow } from 'enzyme';
 
 import path from 'path';
 import fs from 'fs';
 import { name } from '../../../../package.json';
 import * as bundle from '../../..';
 import { size } from '../../Icon';
-import AtlassianIcon from '../../../../glyph/atlassian';
-import BitbucketIcon from '../../../../glyph/bitbucket';
-import ConfluenceIcon from '../../../../glyph/confluence';
+import BookIcon from '../../../../glyph/book';
 import components from '../../../../utils/icons';
 
 // List all files in a directory in Node.js recursively in a synchronous fashion
@@ -55,14 +53,12 @@ describe(name, () => {
         'arrow-right',
         'arrow-up-circle',
         'arrow-up',
-        'atlassian',
         'attachment',
         'audio-circle',
         'audio',
         'backlog',
         'billing-filled',
         'billing',
-        'bitbucket',
         'bitbucket/branches',
         'bitbucket/builds',
         'bitbucket/clone',
@@ -100,7 +96,6 @@ describe(name, () => {
         'code',
         'comment',
         'component',
-        'confluence',
         'copy',
         'creditcard-filled',
         'creditcard',
@@ -198,7 +193,6 @@ describe(name, () => {
         'graph-line',
         'gsuite',
         'highlights',
-        'hipchat',
         'hipchat/audio-only',
         'hipchat/chevron-double-down',
         'hipchat/chevron-double-up',
@@ -219,10 +213,6 @@ describe(name, () => {
         'issue-raise',
         'issue',
         'issues',
-        'jira',
-        'jira-core',
-        'jira-service-desk',
-        'jira-software',
         'jira/blocker',
         'jira/capture',
         'jira/critical',
@@ -318,8 +308,6 @@ describe(name, () => {
         'sign-out',
         'star-filled',
         'star',
-        'statuspage',
-        'stride',
         'subtask',
         'switcher',
         'table',
@@ -337,6 +325,7 @@ describe(name, () => {
         'vid-audio-on',
         'vid-backward',
         'vid-camera-off',
+        'vid-camera-on',
         'vid-connection-circle',
         'vid-forward',
         'vid-full-screen-off',
@@ -348,13 +337,50 @@ describe(name, () => {
         'vid-raised-hand',
         'vid-share-screen',
         'vid-speaking-circle',
+        'vid-volume-full',
+        'vid-volume-half',
+        'vid-volume-muted',
         'video-circle',
         'video-filled',
         'warning',
         'watch-filled',
         'watch',
         'world-small',
+        'star-large',
+        'star-outline',
         'world',
+        'suitcase',
+        'select-clear',
+        'roadmap',
+        'questions',
+        'app-switcher',
+        'media-services/no-image',
+        'label',
+        'media-services/fit-to-page',
+        'media-services/full-screen',
+        'emoji/productivity',
+        'emoji-add',
+        'editor/success',
+        'editor/table-display-options',
+        'editor/media-center',
+        'editor/media-full-width',
+        'editor/media-wide',
+        'editor/media-wrap-left',
+        'editor/media-wrap-right',
+        'editor/layout-three-equal',
+        'editor/layout-three-with-sidebars',
+        'editor/layout-two-equal',
+        'editor/layout-two-left-sidebar',
+        'editor/layout-two-right-sidebar',
+        'editor/background-color',
+        'editor/divider',
+        'editor/horizontal-rule',
+        'department',
+        'child-issues',
+        'chevron-right-large',
+        'chevron-left-large',
+        'check-circle-outline',
+        'like',
       ];
 
       const expectedPaths = expected.map(a => {
@@ -366,7 +392,14 @@ describe(name, () => {
         [],
       ).filter(ab => /.*\.js$/.test(ab));
 
-      expect(actual).toEqual(expect.arrayContaining(expectedPaths));
+      // Additional notes on this check:
+      // We are doing an equality check on the sorted versions of the lists as we want to
+      // ensure that every icon we expect exists, and also that only icons that we expect
+      // exist. This ensures that a new icon cannot be removed without causing this test
+      // to error.
+      // If an icon is Recieved but not expected, it is a new icon we do not expect.
+      // If it is Expected but not received, it means it is an existing icon that has been removed
+      expect(actual.sort()).toEqual(expectedPaths.sort());
       // If you find yourself here and wonder why this list is not auto-generated, then bear in
       // mind that tests are supposed to tell you when a piece of software breaks.
       // As the sole purpose of this component is providing icons:
@@ -407,71 +440,11 @@ describe(name, () => {
     describe('label property', () => {
       it('should accept a label', () => {
         const label = 'my label';
-        const wrapper = mount(<AtlassianIcon label={label} />);
+        const wrapper = mount(<BookIcon label={label} />);
         const span = wrapper.find('span').first();
 
         expect(span.is('[aria-label="my label"]')).toBe(true);
       });
-    });
-  });
-
-  describe('gradients', () => {
-    it('multiple gradients in a single icon should use unique ids', () => {
-      const wrapper = render(<ConfluenceIcon label="My icon" />);
-      // For some reason cheerio will not find linearGradient elements anymore.
-      // Instead we find all the elements inside <defs> and confirm that they are linearGradients
-      const linearGradients = wrapper.find('defs > *');
-
-      expect(linearGradients.length).toBe(2);
-      expect(linearGradients[0].name).toBe('linearGradient');
-      expect(linearGradients[1].name).toBe('linearGradient');
-
-      // now check that they id's are different
-      expect(linearGradients[0].attribs.id).not.toBe(
-        linearGradients[1].attribs.id,
-      );
-    });
-
-    it('ids should be referenced correctly in fill properties', () => {
-      const wrapper = render(<BitbucketIcon label="My icon" />);
-      // for some reason cheerio will no longer find linear gradient elements. Instead we look
-      // inside defs and confirm that we have a linearGradient
-      const gradientEls = wrapper.find('defs > *');
-      expect(gradientEls.length).toBe(1);
-      expect(gradientEls[0].name).toBe('linearGradient');
-
-      // now we can get the ide
-      const gradientId = gradientEls[0].attribs.id;
-      const gradientFillPaths = wrapper
-        .find('path')
-        .filter((i, el) => /^url\(#/.test(el.attribs.fill));
-
-      expect(gradientFillPaths.length).toBeGreaterThanOrEqual(1);
-
-      gradientFillPaths.each((i, el) => {
-        expect(el.attribs.fill).toBe(`url(#${gradientId})`);
-      });
-    });
-
-    it('should have unique ids across icon instances', () => {
-      const icon1 = render(<BitbucketIcon label="My icon" />);
-      const icon2 = render(<BitbucketIcon label="My icon" />);
-      // Again, for some reason cheerio will no longer find linear gradient elements.
-      // Instead we look inside defs and confirm that we have a linearGradient
-      const gradientEls1 = icon1.find('defs > *');
-      expect(gradientEls1.length).toBe(1);
-      expect(gradientEls1[0].name).toBe('linearGradient');
-
-      // and again
-      const gradientEls2 = icon2.find('defs > *');
-      expect(gradientEls2.length).toBe(1);
-      expect(gradientEls2[0].name).toBe('linearGradient');
-
-      // now we can check the id's
-      const gradientId1 = gradientEls1[0].attribs.id;
-      const gradientId2 = gradientEls2[0].attribs.id;
-
-      expect(gradientId1).not.toBe(gradientId2);
     });
   });
 });