+++ packages/core/navigation-next/docs/examples/ui-components/GlobalItem.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import Avatar from '@atlaskit/avatar';
-import AtlassianIcon from '@atlaskit/icon/glyph/atlassian';
+import { AtlassianIcon } from '@atlaskit/logo';
 import SearchIcon from '@atlaskit/icon/glyph/search';
 import HelpIcon from '@atlaskit/icon/glyph/question';
 import { colors } from '@atlaskit/theme';
@@ -20,7 +20,7 @@ export default () => (
       width: '56px',
     }}
   >
-    <GlobalItem icon={() => <AtlassianIcon size="large" />} />
+    <GlobalItem icon={() => <AtlassianIcon size="medium" />} />
     <GlobalItem icon={SearchIcon} />
     <GlobalItem icon={HelpIcon} size="small" />
     <GlobalItem