+++ packages/media/media-store/src/index.ts
@@ -2,6 +2,7 @@ export * from './media-store';
 export * from './models/media';
 export * from './models/auth';
 export * from './upload-controller';
+export * from './models/artifacts';
 
 export {
   uploadFile,