+++ packages/search/global-search/examples/8-Search-in-Confluence-xpsearch.tsx
@@ -27,9 +27,7 @@ export default class extends React.Component {
   render() {
     return (
       <AnalyticsListener onEvent={logEvent} channel="fabric-elements">
-        <GlobalQuickSearchInNavigation
-          useAggregatorForConfluenceObjects={true}
-        />
+        <GlobalQuickSearchInNavigation />
       </AnalyticsListener>
     );
   }