+++ packages/search/global-search/src/__tests__/unit/confluence/ConfluenceQuickSearchContainer.test.tsx
@@ -8,6 +8,7 @@ import { noResultsPeopleSearchClient } from '../mocks/_mockPeopleSearchClient';
 import {
   noResultsConfluenceClient,
   makeConfluenceClient,
+  singleResultQuickNav,
 } from '../mocks/_mockConfluenceClient';
 import { shallowWithIntl } from '../helpers/_intl-enzyme-test-helper';
 import QuickSearchContainer, {
@@ -19,6 +20,7 @@ import { Result } from '../../../model/Result';
 import {
   EMPTY_CROSS_PRODUCT_SEARCH_RESPONSE,
   SearchSession,
+  ABTest,
 } from '../../../api/CrossProductSearchClient';
 import { mockLogger } from '../mocks/_mockLogger';
 
@@ -29,7 +31,7 @@ function render(partialProps?: Partial<Props>) {
     confluenceClient: noResultsConfluenceClient,
     crossProductSearchClient: noResultsCrossProductSearchClient,
     peopleSearchClient: noResultsPeopleSearchClient,
-    useAggregatorForConfluenceObjects: false,
+    useQuickNavForPeopleResults: false,
     useCPUSForPeopleResults: false,
     logger,
     ...partialProps,
@@ -82,6 +84,32 @@ describe('ConfluenceQuickSearchContainer', () => {
     });
   });
 
+  it('should return ab test data', async () => {
+    const abTest: ABTest = {
+      abTestId: 'abTestId',
+      experimentId: 'experimentId',
+      controlId: 'controlId',
+    };
+
+    const wrapper = render({
+      confluenceClient: noResultsConfluenceClient,
+      crossProductSearchClient: {
+        search(query: string) {
+          return Promise.resolve(EMPTY_CROSS_PRODUCT_SEARCH_RESPONSE);
+        },
+        getAbTestData() {
+          return Promise.resolve(abTest);
+        },
+      },
+    });
+    const quickSearchContainer = wrapper.find(QuickSearchContainer);
+    const receivedAbTest = await (quickSearchContainer.props() as QuickSearchContainerProps).getAbTestData(
+      sessionId,
+    );
+
+    expect(receivedAbTest).toMatchObject(abTest);
+  });
+
   it('should return search result', async () => {
     const wrapper = render({
       peopleSearchClient: {
@@ -119,7 +147,6 @@ describe('ConfluenceQuickSearchContainer', () => {
       },
       // assert search performance timings
       timings: {
-        quickNavElapsedMs: expect.any(Number),
         confSearchElapsedMs: expect.any(Number),
         peopleElapsedMs: expect.any(Number),
       },
@@ -143,7 +170,38 @@ describe('ConfluenceQuickSearchContainer', () => {
 
           return Promise.resolve(EMPTY_CROSS_PRODUCT_SEARCH_RESPONSE);
         },
+        getAbTestData(scope: Scope, searchSession: SearchSession) {
+          return Promise.resolve(undefined);
+        },
       },
+    });
+
+    const quickSearchContainer = wrapper.find(QuickSearchContainer);
+    const searchResults = await (quickSearchContainer.props() as QuickSearchContainerProps).getSearchResults(
+      'query',
+      sessionId,
+      100,
+    );
+
+    expect(searchResults.results.people).toEqual([
+      {
+        mentionName: 'mentionName',
+        presenceMessage: 'presenceMessage',
+        analyticsType: 'result-person',
+        resultType: 'person-result',
+        name: 'name',
+        avatarUrl: 'avatarUrl',
+        href: 'href',
+        resultId: expect.any(String),
+      },
+    ]);
+  });
+
+  it('should use quick nav for people results when enabled', async () => {
+    const wrapper = render({
+      useQuickNavForPeopleResults: true,
+      crossProductSearchClient: noResultsCrossProductSearchClient,
+      confluenceClient: singleResultQuickNav(),
     });
 
     const quickSearchContainer = wrapper.find(QuickSearchContainer);