+++ packages/core/calendar/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/calendar
 
+## 7.0.9
+- [patch] Updated dependencies [65c6514](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/65c6514)
+  - @atlaskit/docs@5.0.8
+  - @atlaskit/button@9.0.13
+  - @atlaskit/icon@14.0.0
+
 ## 7.0.8
 - [patch] Fixes bug on next and prev month navigation. [c4770a0](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c4770a0)