+++ packages/core/tree/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @atlaskit/tree
 
+## 4.0.0
+- [major] Implementing nesting and expanding parents while dragging [2bdbf04](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/2bdbf04)
+
+## 3.0.4
+- [patch] Updated dependencies [65c6514](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/65c6514)
+  - @atlaskit/docs@5.0.8
+  - @atlaskit/navigation@33.1.11
+  - @atlaskit/section-message@1.0.8
+  - @atlaskit/icon@14.0.0
+
 ## 3.0.3
 - [patch] Collapsing parent automatically when last child moved out [f4992db](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f4992db)